Forcing the com.google.common packages as optional imports, do not add a generated qualifier
This forces/overrides the import of com.google.common to be optional and
fixes issue #6. bnd adds for changed bundles a hash as qualifier. With
useBndQualifiers false and an empty defaultQualifier, bnd does not
change the name of the generated jars, so they end up with the same name
as in the maven repo